hyprland/ipc: have urgent status of HyprlandWorkspace depending on its clients
Before, the urgent status was manually handled in the event handler.
Now, the HyprlandWorkspace keeps track of its children and watch their
"urgent" status for any change.

Similarly, the HyprlandWorkspace now watches itself on whether it is
focused or not, and removes its urgent mark (and its children) when
focused.